requirements.txt
autopep8==1.3.4
certifi==2018.4.16
chardet==3.0.4
Django==2.0.2
django-cas-ng==3.5.9
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-cleanup==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
extras==1.0.0
gunicorn==19.9.0
idna==2.7
lxml==4.2.4
mysqlclient==1.3.13
pep8==1.7.1
Pillow==6.0.0
pycodestyle==2.3.1
PyJWT==1.5.3
python-cas==1.3.0
pytz==2018.3
requests==2.19.1
six==1.11.0
urllib3==1.23
whitenoise==4.0
